translate-worker/requirements.txt
# HTTP server
flask>=3.0.0

# AI Translation (Hugging Face)
transformers>=4.36.0
torch>=2.0.0
sentencepiece>=0.1.99
accelerate>=0.25.0

# Env
python-dotenv>=1.0.0
